+++ source/MIBRR/R/subroutines.R
@@ -1,7 +1,7 @@
 ### Title:    Subroutines for the MIBRR Package
 ### Author:   Kyle M. Lang
 ### Created:  2017-NOV-28
-### Modified: 2018-MAY-04
+### Modified: 2018-MAY-15
 
 ##--------------------- COPYRIGHT & LICENSING INFORMATION --------------------##
 ##  Copyright (C) 2018 Kyle M. Lang <k.m.lang@uvt.nl>                         ##
@@ -50,16 +50,14 @@ init <- function(doBl,
                          verbose     = verbose,
                          doImp       = doImp,
                          doMcem      = doMcem,
-                         doBl        = doBl)
+                         doBl        = doBl,
+                         seed        = seed)
 
     ## Process and check the user inputs:
     mibrrFit$processInputs()
-
-    ## Setup the PRNG (make sure each target variable gets its own independent
-    ## RNG stream):
-    if(!is.null(seed))
-        .lec.SetPackageSeed(rep(seed, 6))
-    .lec.CreateStream(paste0("mibrrStream", c(1 : length(mibrrFit$targetVars))))
+    
+    ## Setup the PRNG (each target variable gets an independent RNG stream):
+    mibrrFit$setupRng()
     
     ## Store Lambda's prior parameters:
     if(!doMcem) mibrrFit$setLambdaParams(l1 = as.numeric(lam1PriorPar),
@@ -74,25 +72,21 @@ init <- function(doBl,
 
     ## Temporarily fill missing with single imputations:
     if(haveMiss) mibrrFit$simpleImpute(covsOnly = mibrrFit$fimlStarts)
-
-    ## Are we doing any transformations?
-    trans <- mibrrFit$scale || mibrrFit$center
     
     ## Compute summary statistics:
-    if(trans)
-        mibrrFit$computeStats(useFiml = mibrrFit$fimlStarts)
-   
-    if(mibrrFit$center) mibrrFit$meanCenter()
+    mibrrFit$computeStats(useFiml = mibrrFit$fimlStarts)
     
+    if(mibrrFit$center) mibrrFit$meanCenter()
+  
     ## Initialize starting values for the Gibbs sampled parameters.
     ## Important to call this before the NAs are replaced with missCode.
     mibrrFit$startParams()
-    
+  
     ## Fill remaining missing data with an integer code:
     if(mibrrFit$fimlStarts & haveMiss) mibrrFit$applyMissCode()
 
     mibrrFit
-}# END preProcess()
+}# END init()
 
 
 ## Estimate a model using MCEM:
@@ -138,7 +132,7 @@ mcem <- function(mibrrFit) {
     }# END for(i in 1 : totalIters)
     
     mibrrFit
-}# END runMcem()
+}# END mcem()
 
 
 postProcess <- function(mibrrFit) {
@@ -158,8 +152,15 @@ postProcess <- function(mibrrFit) {
     ## Provide some pretty names for the output objects:
     mibrrFit$nameOutput()
 
-    ## Clean up the RNG state:
-    .lec.DeleteStream(paste0("mibrrStream", c(1 : mibrrFit$nTargets)))
+    ## Clean the RNG state:
+    mibrrFit$cleanRng()
+
+    ## Fix rlecuyer's random seed table when we have only 1 remaining stream:
+    check <- !is.null(.lec.Random.seed.table$name) &&
+        !is.matrix(.lec.Random.seed.table$Cg)
     
-mibrrFit
+    if(check) .lec.Random.seed.table[1 : 4] <<-
+                  lapply(.lec.Random.seed.table[1 : 4], matrix, nrow = 1)
+                
+    mibrrFit
 }# END postProcess()