+++ source/MIBRR/src/MibrrGibbs.cpp
@@ -1,7 +1,7 @@
 // Title:    Function definitions for the MibrrGibbs class
 // Author:   Kyle M. Lang
 // Created:  2014-AUG-24
-// Modified: 2018-MAY-04
+// Modified: 2018-MAY-29
 // Purpose:  This class contains the Gibbs sampling-related functions for the
 //           MIBRR package.
 
@@ -58,7 +58,6 @@ int      MibrrGibbs::getNDraws()          const { return _nDraws;              }
 bool     MibrrGibbs::getVerbosity()       const { return _verbose;             }
 bool     MibrrGibbs::getElasticNetFlag()  const { return _useElasticNet;       }
 bool     MibrrGibbs::getDoImp()           const { return _doImp;               }
-bool     MibrrGibbs::getSimpleIntercept() const { return _simpleIntercept;     }
 
 
 VectorXd MibrrGibbs::getLambdas() const
@@ -79,19 +78,18 @@ double MibrrGibbs::getLambdas(int index) const
 //////////////////////////////// MUTATORS //////////////////////////////////////
 
 
-void MibrrGibbs::setBetas      (VectorXd &betas)   { _betas = betas;           }
-void MibrrGibbs::setTaus       (ArrayXd &taus)     { _taus = taus;             }
-void MibrrGibbs::setSigma      (double sigma)      { _sigma = sigma;           }
-void MibrrGibbs::setTargetIndex(int index)         { _targetIndex = index;     }
-void MibrrGibbs::setNDraws     (int nDraws)        { _nDraws = nDraws;         }
-void MibrrGibbs::setDoImp      (bool doImp)        { _doImp = doImp;           }
-void MibrrGibbs::beQuiet       ()                  { _verbose = false;         }
+void MibrrGibbs::setBetas      (VectorXd &betas)   { _betas         = betas;   }
+void MibrrGibbs::setTaus       (ArrayXd &taus)     { _taus          = taus;    }
+void MibrrGibbs::setSigma      (double sigma)      { _sigma         = sigma;   }
+void MibrrGibbs::setTargetIndex(int index)         { _targetIndex   = index;   }
+void MibrrGibbs::setNDraws     (int nDraws)        { _nDraws        = nDraws;  }
+void MibrrGibbs::setDoImp      (bool doImp)        { _doImp         = doImp;   }
+void MibrrGibbs::beQuiet       ()                  { _verbose       = false;   }
 void MibrrGibbs::doBl          ()                  { _useElasticNet = false;   }
-void MibrrGibbs::useSimpleInt  ()                  { _simpleIntercept = true;  }
-void MibrrGibbs::doFullBayes   ()                  { _fullBayes = true;        }
-void MibrrGibbs::setLam1Parms  (VectorXd& l1Parms) { _l1Parms = l1Parms;       }
-void MibrrGibbs::setLam2Parms  (VectorXd& l2Parms) { _l2Parms = l2Parms;       }
-void MibrrGibbs::setLambdas    (VectorXd& lambdas) { _lambdas = lambdas;       }
+void MibrrGibbs::doFullBayes   ()                  { _fullBayes     = true;    }
+void MibrrGibbs::setLam1Parms  (VectorXd& l1Parms) { _l1Parms       = l1Parms; }
+void MibrrGibbs::setLam2Parms  (VectorXd& l2Parms) { _l2Parms       = l2Parms; }
+void MibrrGibbs::setLambdas    (VectorXd& lambdas) { _lambdas       = lambdas; }
 
 
 void MibrrGibbs::setLambdas(double lambda1, double lambda2) 
@@ -121,11 +119,11 @@ void MibrrGibbs::startParameters(VectorXd &betaStarts,
 				 double   lambda1Start,
 				 double   lambda2Start)
 {
-  _betas         = betaStarts;
-  _taus          = tauStarts;
-  _sigma         = sigmaStart;
-  _lambdas[0]    = lambda1Start;
-  _lambdas[1]    = lambda2Start;
+  _betas      = betaStarts;
+  _taus       = tauStarts;
+  _sigma      = sigmaStart;
+  _lambdas[0] = lambda1Start;
+  _lambdas[1] = lambda2Start;
 }
 
 
@@ -134,10 +132,10 @@ void MibrrGibbs::startParameters(VectorXd &betaStarts,
 				 double   sigmaStart,
 				 VectorXd &lambdaStartVec)
 {
-  _betas         = betaStarts;
-  _taus          = tauStarts;
-  _sigma         = sigmaStart;
-  _lambdas       = lambdaStartVec;
+  _betas   = betaStarts;
+  _taus    = tauStarts;
+  _sigma   = sigmaStart;
+  _lambdas = lambdaStartVec;
 }
 
 
@@ -182,15 +180,13 @@ void MibrrGibbs::updateLambdas(const MibrrData &mibrrData)
   else {
     double lam2 = _lambdas[1];
     
-    // sample new value of lambda1:
+    // Sample new value of lambda1:
     double shape = _l1Parms[0] + nPreds / 2.0;
     double rate  = _l1Parms[1] + tauSum / (8.0 * lam2 * _sigma);
     
     _lambdas[0] = sqrt(drawGamma(shape, rate));
     
-    // sample new value of lambda2:
-    //double fBeta = _betas.tail(nPreds).asArray().square().sum(); 
-    
+    // Sample new value of lambda2:
     double chi = _l2Parms[0] + (pow(lam1, 2) * tauSum) / (4 * _sigma);
     double psi =
       _l2Parms[1] + ((tauSum / (_taus - 1.0).sum()) *
@@ -211,7 +207,7 @@ void MibrrGibbs::updateTaus(const MibrrData &mibrrData)
   ArrayXd tauMeans;
   
   try {
-    if(_useElasticNet) {// Miben Version
+    if(_useElasticNet) {// MIBEN Version
       ArrayXd tauMeansNumerator = ArrayXd::Constant(nPreds, _lambdas[0]);
       
       tauMeans = tauMeansNumerator.sqrt() /
@@ -251,76 +247,72 @@ void MibrrGibbs::updateTaus(const MibrrData &mibrrData)
 
 void MibrrGibbs::updateBetas(const MibrrData &mibrrData)
 {
-  int             nPreds = mibrrData.nPreds();
-  int             nObs   = mibrrData.nResp(_targetIndex);
-  MatrixXd        tmpMat;
-  LDLT <MatrixXd> aMatrixCholesky;
+  int             nPreds    = mibrrData.nPreds();
+  int             nObs      = mibrrData.nResp(_targetIndex);
+  VectorXd        transTaus = VectorXd::Zero(nPreds);
+  MatrixXd        penalty;
+  LDLT <MatrixXd> aMatChol;
   
   if(_useElasticNet) {// MIBEN Version
-    VectorXd transformedTaus = _taus / (_taus - 1.0);
-    tmpMat = _lambdas[1] * transformedTaus.asDiagonal();    
+    transTaus = _taus / (_taus - 1.0);
+    penalty   = _lambdas[1] * transTaus.asDiagonal();    
   }
-  else {// MIBL Version
-    VectorXd transformedTaus = 1.0 / _taus;
-    tmpMat = transformedTaus.asDiagonal();
+  else {              // MIBL Version
+    transTaus = 1.0 / _taus;
+    penalty   = transTaus.asDiagonal();
   }
-  
-  MatrixXd aMatrix;
-  aMatrix = mibrrData.getIVs(_targetIndex, true).transpose() *
-    mibrrData.getIVs(_targetIndex, true) + tmpMat;
+    
+  MatrixXd aMat = mibrrData.getIVs(_targetIndex, true).transpose() *
+    mibrrData.getIVs(_targetIndex, true) + penalty;
   
   VectorXd betaMeans;
-  MatrixXd betaCovariances;
+  MatrixXd betaCov;
   try {
     // Compute the Cholesky decomposition of the IV's penalized crossproducts,
     // and use it to compute the moments of beta's fully conditional posterior:
-    aMatrixCholesky.compute(aMatrix);
+    aMatChol.compute(aMat);
     betaMeans =
-      aMatrixCholesky.solve(mibrrData.getIVs(_targetIndex, true).transpose() *
-			    mibrrData.getDV(_targetIndex)); 
-    
-    tmpMat = _sigma * MatrixXd::Identity(nPreds, nPreds);   
-    betaCovariances = aMatrixCholesky.solve(tmpMat);
+      aMatChol.solve(mibrrData.getIVs(_targetIndex, true).transpose() *
+		     mibrrData.getDV(_targetIndex)); 
+    betaCov =
+      aMatChol.solve(_sigma * MatrixXd::Identity(nPreds, nPreds));
   }
   catch(exception &e) { betaError(e); }
-
-  // Draw a new value of the intercept term:
-  double intSd = sqrt(_sigma / double(nObs));
-  double intMean;
-  if(_simpleIntercept)
-    intMean = mibrrData.getDV(_targetIndex).mean();
-  else
-    intMean = mibrrData.getDV(_targetIndex).mean() -
-      mibrrData.getIVs(_targetIndex, true).colwise().mean() *
-      _betas.tail(nPreds);
-  
-  VectorXd newBetas(nPreds + 1);  
-  newBetas[0] = drawNorm(intMean, intSd);
   
   // Draw new values of the regression slope coefficients:
-  newBetas.tail(nPreds) = drawMvn(betaMeans, betaCovariances);
+  VectorXd newBetas(nPreds + 1); 
+  newBetas.tail(nPreds) = drawMvn(betaMeans, betaCov);
+  
+  // Compute parameters of the intercept's distribution:
+  double intSd   = sqrt(_sigma / double(nObs));
+  double intMean = mibrrData.getDV(_targetIndex).mean() -
+    mibrrData.getIVs(_targetIndex, true).colwise().mean() *
+    newBetas.tail(nPreds);
+  
+  // Draw a new value of the intercept term:  
+  newBetas[0] = drawNorm(intMean, intSd);
   
   _betas = newBetas; // Store the updated Betas
   
   // Add the updated Betas to their Gibbs sample:
   if(_storeGibbsSamples) _betaSam.row(_drawNum) = newBetas.transpose();
 }// END updateBetas()
-    
+
 
 
 void MibrrGibbs::updateSigma(const MibrrData &mibrrData)
 {
-  double   newSigma;
-  int      nPreds        = mibrrData.nPreds();
-  int      nObs          = mibrrData.nResp(_targetIndex);
-  VectorXd tmpBiasVector = VectorXd::Ones(nObs);
+  double newSigma;
+  int    nPreds = mibrrData.nPreds();
+  int    nObs   = mibrrData.nResp(_targetIndex);
+  
+  // Compute the linear predictor:
+  VectorXd eta = _betas[0] * VectorXd::Ones(nObs) +
+    mibrrData.getIVs(_targetIndex, true) * _betas.tail(nPreds);
   
   // Compute the regularized residual sum of squares:
-  double sse =
-    (mibrrData.getDV(_targetIndex) - _betas[0] * tmpBiasVector -
-     mibrrData.getIVs(_targetIndex, true) * _betas.tail(nPreds)).transpose() *
-    (mibrrData.getDV(_targetIndex) - _betas[0] * tmpBiasVector -
-     mibrrData.getIVs(_targetIndex, true) * _betas.tail(nPreds));
+  double sse = (mibrrData.getDV(_targetIndex) - eta).transpose() *
+    (mibrrData.getDV(_targetIndex) - eta);
   
   if(_useElasticNet) {// MIBEN Version
     double scaleSum =
@@ -331,27 +323,30 @@ void MibrrGibbs::updateSigma(const MibrrData &mibrrData)
       0.5 * (sse + _lambdas[1] * scaleSum +
 	     (pow(_lambdas[0], 2) / (4.0 * _lambdas[1])) * _taus.sum());
     
-    bool   isDrawValid = false;
+    bool   validDraw = false;
     double testDraw;
-    while(!isDrawValid) {// Rejection sampling to draw a Sigma variate
-      testDraw         = drawInvGamma(sigmaShape, sigmaScale);
-      double threshold = _unif(_gen);
-      double igShape   = pow(_lambdas[0], 2) / (8.0 * testDraw * _lambdas[1]);
-      double igDraw    = calcIncGamma(0.5, igShape, false);
-      isDrawValid      = log(threshold) <=
+    while(!validDraw) {// Rejection sampling to draw a Sigma variate
+      testDraw = drawInvGamma(sigmaShape, sigmaScale);
+      
+      double igShape = pow(_lambdas[0], 2) / (8.0 * testDraw * _lambdas[1]);
+      double igDraw  = calcIncGamma(0.5, igShape, false);
+      
+      double thresh = _unif(_gen);
+      validDraw     = log(thresh) <=
 	(double(nPreds) * log(tgamma(0.5))) - (double(nPreds) * log(igDraw));
+      
       Rcpp::checkUserInterrupt();
     };
     newSigma = testDraw;
   }
   else {// MIBL Version
-    VectorXd transformedTaus = 1.0 / _taus;
-    MatrixXd tmpMat = transformedTaus.asDiagonal();
+    VectorXd transTaus = 1.0 / _taus;
+    MatrixXd tmpMat    = transTaus.asDiagonal();
     
     double penaltyTerm =
       _betas.tail(nPreds).transpose() * tmpMat * _betas.tail(nPreds);
-    double sigmaShape = 0.5 * ((double(nObs) - 1.0) + double(nPreds));
-    double sigmaScale = 0.5 * (sse + penaltyTerm);
+    double sigmaShape  = 0.5 * ((double(nObs) - 1.0) + double(nPreds));
+    double sigmaScale  = 0.5 * (sse + penaltyTerm);
     
     // Draw a new sigma variate:
     newSigma = drawInvGamma(sigmaShape, sigmaScale);
@@ -367,24 +362,20 @@ void MibrrGibbs::updateSigma(const MibrrData &mibrrData)
 
 void MibrrGibbs::updateImputations(MibrrData &mibrrData)
 {
-  int         nMiss         = mibrrData.nMiss(_targetIndex);
-  int         nPreds        = mibrrData.nPreds();
-  VectorXd    tmpBiasVector = VectorXd::Ones(nMiss);
-  VectorXd    errorVector(nMiss);
-  
-  // Draw the residual error terms for the imputation model:
-  for(int i = 0; i < nMiss; i++) errorVector[i] = drawNorm(0.0, sqrt(_sigma));
-  
+  int nMiss  = mibrrData.nMiss(_targetIndex);
+  int nPreds = mibrrData.nPreds();
+   
   // Draw a vector of imputations from the posterior predictive distribution
   // of the missing data:
-  VectorXd tmpImps = _betas[0] * tmpBiasVector +
-    mibrrData.getIVs(_targetIndex, false) * _betas.tail(nPreds) + errorVector;
+  VectorXd imps = _betas[0] * VectorXd::Ones(nMiss) +
+    mibrrData.getIVs(_targetIndex, false) * _betas.tail(nPreds) +
+    drawNorm(nMiss, 0.0, sqrt(_sigma));
   
   // Replace the missing data in the target variable with the imputations:
-  mibrrData.fillMissing(tmpImps, _targetIndex);
+  mibrrData.fillMissing(imps, _targetIndex);
   
   // Add the updated imputations to their Gibbs sample:
-  if(_storeGibbsSamples) _impSam.row(_drawNum) = tmpImps.transpose();
+  if(_storeGibbsSamples) _impSam.row(_drawNum) = imps.transpose();
 }// END updateImputations()